3681.   [port]          Update the Windows build system to support feature                         selection and WIN64 builds.  This is a work in                         progress. [RT #34160]

+++ lib/dns/master.c
@@ -702,7 +702,7 @@ genname(char *name, int it, char *buffer, size_t length) {
 	isc_boolean_t nibblemode;
 
 	r.base = buffer;
-	r.length = length;
+	r.length = (unsigned int)length;
 
 	while (*name != '\0') {
 		if (*name == '$') {
@@ -2088,7 +2088,7 @@ read_and_check(isc_boolean_t do_read, isc_buffer_t *buffer,
 					f, NULL);
 		if (result != ISC_R_SUCCESS)
 			return (result);
-		isc_buffer_add(buffer, len);
+		isc_buffer_add(buffer, (unsigned int)len);
 	} else if (isc_buffer_remaininglength(buffer) < len)
 		return (ISC_R_RANGE);
 
@@ -2125,7 +2125,7 @@ load_header(dns_loadctx_t *lctx) {
 		return (result);
 	}
 
-	isc_buffer_add(&target, commonlen);
+	isc_buffer_add(&target, (unsigned int)commonlen);
 	header.format = isc_buffer_getuint32(&target);
 	if (header.format != lctx->format) {
 		(*callbacks->error)(callbacks, "dns_master_load: "
@@ -2160,7 +2160,7 @@ load_header(dns_loadctx_t *lctx) {
 		return (result);
 	}
 
-	isc_buffer_add(&target, remainder);
+	isc_buffer_add(&target, (unsigned int)remainder);
 	header.dumptime = isc_buffer_getuint32(&target);
 	if (header.version == DNS_RAWFORMAT_VERSION) {
 		header.flags = isc_buffer_getuint32(&target);
@@ -2346,7 +2346,7 @@ load_raw(dns_loadctx_t *lctx) {
 					lctx->f, NULL);
 		if (result != ISC_R_SUCCESS)
 			goto cleanup;
-		isc_buffer_add(&target, readlen);
+		isc_buffer_add(&target, (unsigned int)readlen);
 
 		/* Construct RRset headers */
 		rdatalist.rdclass = isc_buffer_getuint16(&target);
