+++ lib/isc/win32/file.c
@@ -333,7 +333,7 @@ isc_file_template(const char *path, const char *templet, char *buf,
 	s = strrchr(path, '\\');
 
 	if (s != NULL) {
-		if ((s - path + 1 + strlen(templet) + 1) > buflen)
+	  if ((s - path + 1 + strlen(templet) + 1) > (ssize_t)buflen)
 			return (ISC_R_NOSPACE);
 
 		strncpy(buf, path, s - path + 1);
@@ -628,7 +628,7 @@ isc_file_absolutepath(const char *filename, char *path, size_t pathlen) {
 	REQUIRE(filename != NULL);
 	REQUIRE(path != NULL);
 
-	retval = GetFullPathName(filename, pathlen, path, &ptrname);
+	retval = GetFullPathName(filename, (DWORD) pathlen, path, &ptrname);
 
 	/* Something went wrong in getting the path */
 	if (retval == 0)
@@ -751,7 +751,7 @@ isc_file_mmap(void *addr, size_t len, int prot,
 		len = end - offset;
 
 	buf = malloc(len);
-	ret = read(fd, buf, len);
+	ret = read(fd, buf, (unsigned int) len);
 	if (ret != (ssize_t) len) {
 		free(buf);
 		buf = NULL;
